+++ packages/applications/web/src/bootstrap.ts
@@ -48,7 +48,7 @@ export const bootstrap = async (): Promise<UnsetupApp> => {
         téléverserFichierDossierRaccordementAdapter,
     },
     projet: { téléverserFichier: téléverserFichierAdapter },
-    dépôtGarantiesFinancières: { téléverserFichier: téléverserFichierAdapter },
+    garantiesFinancières: { téléverserFichier: téléverserFichierAdapter },
   });
 
   const unsetupDomainViews = await setupDomainViews({
