+++ packages/domain/src/domain.setup.ts
@@ -3,7 +3,7 @@ import { setupGestionnaireRéseau } from './gestionnaireRéseau/gestionnaireRés
 import { ProjetDependencies, setupProjet } from './projet/projet.setup';
 import { RaccordementDependencies, setupRaccordement } from './raccordement/raccordement.setup';
 import {
-  DépôtGarantiesFinancièresDependencies,
+  GarantiesFinancièresDependencies,
   setupDépôtGarantiesFinancières,
 } from './garantiesFinancières/garantiesFinancières.setup';
 
@@ -15,10 +15,7 @@ export type DomainDependencies = {
   };
   raccordement: Omit<RaccordementDependencies, keyof DomainDependencies['common']>;
   projet: Omit<ProjetDependencies, keyof DomainDependencies['common']>;
-  dépôtGarantiesFinancières: Omit<
-    DépôtGarantiesFinancièresDependencies,
-    keyof DomainDependencies['common']
-  >;
+  garantiesFinancières: Omit<GarantiesFinancièresDependencies, keyof DomainDependencies['common']>;
 };
 
 export type UnsetupDomain = () => Promise<void>;
@@ -27,7 +24,7 @@ export const setupDomain = async ({
   common,
   raccordement,
   projet,
-  dépôtGarantiesFinancières,
+  garantiesFinancières,
 }: DomainDependencies): Promise<UnsetupDomain> => {
   setupRaccordement({
     ...common,
@@ -36,7 +33,7 @@ export const setupDomain = async ({
 
   setupGestionnaireRéseau(common);
 
-  setupDépôtGarantiesFinancières({ ...common, ...dépôtGarantiesFinancières });
+  setupDépôtGarantiesFinancières({ ...common, ...garantiesFinancières });
 
   const unsubscribeProjet = await setupProjet({ ...common, ...projet });
