+++ packages/domain/src/domain.usecase.ts
@@ -1,4 +1,4 @@
-import { DépôtGarantiesFinancièresUseCase } from './garantiesFinancières/garantiesFinancières.usecase';
+import { GarantiesFinancièresUseCase } from './garantiesFinancières/garantiesFinancières.usecase';
 import { GestionnaireRéseauUseCase } from './gestionnaireRéseau/gestionnaireRéseau.usecase';
 import { ProjetUseCase } from './projet/projet.usecase';
 import { RaccordementUsecase } from './raccordement/raccordement.usecase';
@@ -7,4 +7,4 @@ export type DomainUseCase =
   | GestionnaireRéseauUseCase
   | ProjetUseCase
   | RaccordementUsecase
-  | DépôtGarantiesFinancièresUseCase;
+  | GarantiesFinancièresUseCase;
