+++ packages/domain/src/garantiesFinancières/actuelles/enregistrementGarantiesFinancières.event.ts
