+++ packages/domain/src/garantiesFinancières/actuelles/enregistrerAttestationGarantiesFinancières.command.ts
@@ -1,10 +1,10 @@
 import { Message, MessageHandler, mediator } from 'mediateur';
-import { IdentifiantProjetValueType } from '../projet.valueType';
-import { AttestationConstitution } from '../../garantiesFinancières/garantiesFinancières.valueType';
+import { IdentifiantProjetValueType } from '../../projet/projet.valueType';
+import { AttestationConstitution } from '../garantiesFinancières.valueType';
 import { Publish } from '@potentiel/core-domain';
-import { createProjetAggregateId } from '../projet.aggregate';
-import { verifyGarantiesFinancièresAttestationForCommand } from './verifyGarantiesFinancièresAttestationForCommand';
-import { AttestationGarantiesFinancièresEnregistréeEvent } from './garantiesFinancières.event';
+import { createProjetAggregateId } from '../../projet/projet.aggregate';
+import { verifyGarantiesFinancièresAttestationForCommand } from '../verifyGarantiesFinancièresAttestationForCommand';
+import { AttestationGarantiesFinancièresEnregistréeEvent } from './enregistrementGarantiesFinancières.event';
 import { TéléverserFichierPort } from '../../common.ports';
 
 export type EnregistrerAttestationGarantiesFinancièresCommand = Message<
