+++ packages/domain/src/garantiesFinancières/garantiesFinancières.setup.ts
@@ -1,4 +1,14 @@
 import {
+  EnregistrerAttestationGarantiesFinancièresDependencies,
+  registerEnregistrerAttestationGarantiesFinancièresCommand,
+} from './actuelles/enregistrerAttestationGarantiesFinancières.command';
+import { registerEnregistrerGarantiesFinancièresUseCase } from './actuelles/enregistrerGarantiesFinancières.usecase';
+import { registerEnregistrerGarantiesFinancièresComplètesCommand } from './actuelles/enregistrerGarantiesFinancièresComplètes.command';
+import {
+  EnregistrerTypeGarantiesFinancièresDependencies,
+  registerEnregistrerTypeGarantiesFinancièresCommand,
+} from './actuelles/enregistrerTypeGarantiesFinancières.command';
+import {
   DéposerGarantiesFinancièresDependencies,
   registerDéposerGarantiesFinancièresCommand,
 } from './dépôt/déposerGarantiesFinancières.command';
@@ -9,17 +19,23 @@ import {
 } from './dépôt/modifierDépôtGarantiesFinancières.command';
 import { registerModifierDépôtGarantiesFinancièresUseCase } from './dépôt/modifierdépôtGarantiesFinancières.usecase';
 
-export type DépôtGarantiesFinancièresDependencies = DéposerGarantiesFinancièresDependencies &
-  ModifierDépôtGarantiesFinancièresDependencies;
+export type GarantiesFinancièresDependencies = DéposerGarantiesFinancièresDependencies &
+  ModifierDépôtGarantiesFinancièresDependencies &
+  EnregistrerTypeGarantiesFinancièresDependencies &
+  EnregistrerAttestationGarantiesFinancièresDependencies;
 
 export const setupDépôtGarantiesFinancières = async (
-  dependencies: DépôtGarantiesFinancièresDependencies,
+  dependencies: GarantiesFinancièresDependencies,
 ) => {
   // commands
   registerDéposerGarantiesFinancièresCommand(dependencies);
   registerModifierDépôtGarantiesFinancièresCommand(dependencies);
+  registerEnregistrerTypeGarantiesFinancièresCommand(dependencies);
+  registerEnregistrerAttestationGarantiesFinancièresCommand(dependencies);
+  registerEnregistrerGarantiesFinancièresComplètesCommand(dependencies);
 
   // usecases
   registerDéposerGarantiesFinancièresUseCase();
   registerModifierDépôtGarantiesFinancièresUseCase();
+  registerEnregistrerGarantiesFinancièresUseCase();
 };
