+++ packages/domain/src/garantiesFinancières/verifyGarantiesFinancièresAttestationForCommand.ts
@@ -1,0 +1,11 @@
+import { DateConstitutionGarantiesFinancièreDansLeFuturErreur } from './garantiesFinancières.error';
+import { AttestationConstitution } from './garantiesFinancières.valueType';
+
+export const verifyGarantiesFinancièresAttestationForCommand = ({
+  date,
+}: AttestationConstitution) => {
+  if (date.estDansLeFutur()) {
+    throw new DateConstitutionGarantiesFinancièreDansLeFuturErreur();
+  }
+  return;
+};
