--- packages/domain/src/projet/garantiesFinancières/verifyGarantiesFinancièresAttestationForCommand.ts
@@ -1,11 +1,0 @@
-import { DateConstitutionGarantiesFinancièreDansLeFuturErreur } from '../../garantiesFinancières/garantiesFinancières.error';
-import { AttestationConstitution } from '../../garantiesFinancières/garantiesFinancières.valueType';
-
-export const verifyGarantiesFinancièresAttestationForCommand = ({
-  date,
-}: AttestationConstitution) => {
-  if (date.estDansLeFutur()) {
-    throw new DateConstitutionGarantiesFinancièreDansLeFuturErreur();
-  }
-  return;
-};
