+++ packages/domain/src/projet/projet.command.ts
@@ -1,12 +1,6 @@
-import { EnregistrerAttestationGarantiesFinancièresCommand } from './garantiesFinancières/enregistrerAttestationGarantiesFinancières.command';
-import { EnregistrerGarantiesFinancièresComplètesCommand } from './garantiesFinancières/enregistrerGarantiesFinancièresComplètes.command';
-import { EnregistrerTypeGarantiesFinancièresCommand } from './garantiesFinancières/enregistrerTypeGarantiesFinancières.command';
 import { DéclarerGestionnaireRéseauProjetCommand } from './gestionnaireRéseau/déclarer/déclarerGestionnaireRéseauProjet.command';
 import { ModifierGestionnaireRéseauProjetCommand } from './gestionnaireRéseau/modifier/modifierGestionnaireRéseauProjet.command';
 
 export type ProjetCommand =
   | DéclarerGestionnaireRéseauProjetCommand
-  | ModifierGestionnaireRéseauProjetCommand
-  | EnregistrerTypeGarantiesFinancièresCommand
-  | EnregistrerAttestationGarantiesFinancièresCommand
-  | EnregistrerGarantiesFinancièresComplètesCommand;
+  | ModifierGestionnaireRéseauProjetCommand;
