+++ packages/domain/src/projet/projet.event.ts
@@ -1,9 +1,4 @@
 import {
-  AttestationGarantiesFinancièresEnregistréeEvent,
-  TypeGarantiesFinancièresEnregistréEventV1,
-  TypeGarantiesFinancièresEnregistréEventV0,
-} from './garantiesFinancières/garantiesFinancières.event';
-import {
   GestionnaireRéseauProjetModifiéEvent,
   GestionnaireRéseauProjetDéclaréEvent,
 } from './gestionnaireRéseau/gestionnaireRéseau.event';
@@ -12,9 +7,4 @@ type GestionnaireRéseauProjetEvent =
   | GestionnaireRéseauProjetModifiéEvent
   | GestionnaireRéseauProjetDéclaréEvent;
 
-export type GarantiesFinancièresEvent =
-  | AttestationGarantiesFinancièresEnregistréeEvent
-  | TypeGarantiesFinancièresEnregistréEventV1
-  | TypeGarantiesFinancièresEnregistréEventV0;
-
-export type ProjetEvent = GestionnaireRéseauProjetEvent | GarantiesFinancièresEvent;
+export type ProjetEvent = GestionnaireRéseauProjetEvent;
