+++ packages/domain/src/projet/projet.usecase.ts
@@ -1,6 +1,3 @@
-import { EnregistrerGarantiesFinancièresUseCase } from './garantiesFinancières/enregistrerGarantiesFinancières.usecase';
 import { ModifierGestionnaireRéseauProjetUseCase } from './gestionnaireRéseau/modifier/modifierGestionnaireRéseauProjet.usecase';
 
-export type ProjetUseCase =
-  | ModifierGestionnaireRéseauProjetUseCase
-  | EnregistrerGarantiesFinancièresUseCase;
+export type ProjetUseCase = ModifierGestionnaireRéseauProjetUseCase;
