+++ packages/domain/src/projet/projet.valueType.ts
@@ -1,7 +1,5 @@
 import { Option, isNone, none } from '@potentiel/monads';
 
-export * from '../garantiesFinancières/garantiesFinancières.valueType';
-
 export type RawIdentifiantProjet = `${string}#${string}#${string}#${string}`;
 
 export type IdentifiantProjet = {
